Remove notes app and move codes to shipments
