--- apps/notes/__init__.py
@@ -1,5 +1,0 @@
-from django.apps import AppConfig
-
-
-class NotesConfig(AppConfig):
-    name = 'apps.notes'
