+++ apps/shipments/serializers/__init__.py
@@ -5,3 +5,4 @@ from .shipment_history import ChangesDiffSerializer
 from .tracking_data import TrackingDataSerializer, TrackingDataToDbSerializer, UnvalidatedTrackingDataSerializer
 from .shipment_overview import DevicesQueryParamsSerializer
 from .shipment_action import ShipmentActionRequestSerializer, ActionType
+from .note import ShipmentNoteSerializer, ShipmentNoteCreateSerializer
