+++ apps/shipments/views/__init__.py
@@ -4,3 +4,4 @@ from .tracking_data import DeviceViewSet
 from .history import ShipmentHistoryListView
 from .shipment_overview import ListDevicesStatus
 from .shipment_action import ShipmentActionsView
+from .note import ShipmentNoteViewSet
