+++ conf/base.py
@@ -130,7 +130,6 @@ INSTALLED_APPS = [
     'apps.schema',
     'apps.documents',
     'apps.imports',
-    'apps.notes',
     'channels',
     'simple_history',
     'rangefilter',
